Refactor/rule

+++ SmartHomeManager/SmartHomeManager.View/src/components/Rules/Buttons.js
@@ -2,30 +2,40 @@ import { Button, Td } from "@chakra-ui/react";
 import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import axios from "axios";
-function Buttons({props, deleteRule}) {
+
+import ConfirmationScreen from "../../pages/rules/ConfirmationScreen";
+
+function Buttons({props, deleteRule,editButton}) {
 	const [isDeleted, setIsDeleted] = useState(false);
+	const [showConfirmation, setShowConfirmation] = useState(false);
 	console.log(props);
 	return (
 		<Td>
 			<Button ml={2} colorScheme="blue">
-				<Link
-					to={`/scenario/edit/${props.ruleId}`}
-					state={props.props}
-				>
-					Edit
-				</Link>
+				{editButton({props})}
 			</Button>
 			<Button
 				ml={2}
 				colorScheme="red"
 				onClick={() => {
-					deleteRule(props.ruleId).catch((error) => {
-						console.error(error);
-					});
+					setShowConfirmation(true);
 				}}
 			>
 				Delete
 			</Button>
+			{showConfirmation && (
+				<ConfirmationScreen
+					onConfirm={() => {
+						deleteRule(props.ruleId).catch((error) => {
+							console.error(error);
+							});
+						setShowConfirmation(false);
+					}}
+					Close={()=>{
+						setShowConfirmation(false);
+					}}
+				/>
+			)}
 		</Td>
 	);
 }
